Final merge of VC4 mesa's with Vulkan source

+++ mesa 3D driver/src/util/hash_table.c
@@ -191,26 +191,6 @@ _mesa_hash_table_create(void *mem_ctx,
    return ht;
 }
 
-static uint32_t
-key_u32_hash(const void *key)
-{
-   uint32_t u = (uint32_t)(uintptr_t)key;
-   return _mesa_hash_uint(&u);
-}
-
-static bool
-key_u32_equals(const void *a, const void *b)
-{
-   return (uint32_t)(uintptr_t)a == (uint32_t)(uintptr_t)b;
-}
-
-/* key == 0 and key == deleted_key are not allowed */
-struct hash_table *
-_mesa_hash_table_create_u32_keys(void *mem_ctx)
-{
-   return _mesa_hash_table_create(mem_ctx, key_u32_hash, key_u32_equals);
-}
-
 struct hash_table *
 _mesa_hash_table_clone(struct hash_table *src, void *dst_mem_ctx)
 {
@@ -254,13 +234,6 @@ _mesa_hash_table_destroy(struct hash_table *ht,
    ralloc_free(ht);
 }
 
-static void
-hash_table_clear_fast(struct hash_table *ht)
-{
-   memset(ht->table, 0, sizeof(struct hash_entry) * hash_sizes[ht->size_index].size);
-   ht->entries = ht->deleted_entries = 0;
-}
-
 /**
  * Deletes all entries of the given hash table without deleting the table
  * itself or changing its structure.
@@ -271,22 +244,20 @@ void
 _mesa_hash_table_clear(struct hash_table *ht,
                        void (*delete_function)(struct hash_entry *entry))
 {
-   if (!ht)
-      return;
-
    struct hash_entry *entry;
 
-   if (delete_function) {
-      for (entry = ht->table; entry != ht->table + ht->size; entry++) {
-         if (entry_is_present(ht, entry))
-            delete_function(entry);
+   for (entry = ht->table; entry != ht->table + ht->size; entry++) {
+      if (entry->key == NULL)
+         continue;
 
-         entry->key = NULL;
-      }
-      ht->entries = 0;
-      ht->deleted_entries = 0;
-   } else
-      hash_table_clear_fast(ht);
+      if (delete_function != NULL && entry->key != ht->deleted_key)
+         delete_function(entry);
+
+      entry->key = NULL;
+   }
+
+   ht->entries = 0;
+   ht->deleted_entries = 0;
 }
 
 /** Sets the value of the key pointer used for deleted entries in the table.
@@ -390,12 +361,6 @@ _mesa_hash_table_rehash(struct hash_table *ht, unsigned new_size_index)
 {
    struct hash_table old_ht;
    struct hash_entry *table;
-
-   if (ht->size_index == new_size_index && ht->deleted_entries == ht->max_entries) {
-      hash_table_clear_fast(ht);
-      assert(!ht->entries);
-      return;
-   }
 
    if (new_size_index >= ARRAY_SIZE(hash_sizes))
       return;
@@ -545,27 +510,6 @@ void _mesa_hash_table_remove_key(struct hash_table *ht,
 }
 
 /**
- * This function is an iterator over the hash_table when no deleted entries are present.
- *
- * Pass in NULL for the first entry, as in the start of a for loop.
- */
-struct hash_entry *
-_mesa_hash_table_next_entry_unsafe(const struct hash_table *ht, struct hash_entry *entry)
-{
-   assert(!ht->deleted_entries);
-   if (!ht->entries)
-      return NULL;
-   if (entry == NULL)
-      entry = ht->table;
-   else
-      entry = entry + 1;
-   if (entry != ht->table + ht->size)
-      return entry->key ? entry : _mesa_hash_table_next_entry_unsafe(ht, entry);
-
-   return NULL;
-}
-
-/**
  * This function is an iterator over the hash table.
  *
  * Pass in NULL for the first entry, as in the start of a for loop.  Note that
@@ -629,12 +573,6 @@ uint32_t
 _mesa_hash_data(const void *data, size_t size)
 {
    return XXH32(data, size, 0);
-}
-
-uint32_t
-_mesa_hash_data_with_seed(const void *data, size_t size, uint32_t seed)
-{
-   return XXH32(data, size, seed);
 }
 
 uint32_t
@@ -788,35 +726,65 @@ _mesa_hash_table_u64_create(void *mem_ctx)
    return ht;
 }
 
-static void
-_mesa_hash_table_u64_delete_key(struct hash_entry *entry)
-{
-   if (sizeof(void *) == 8)
-      return;
-
-   struct hash_key_u64 *_key = (struct hash_key_u64 *)entry->key;
-
-   if (_key)
-      free(_key);
-}
-
 void
-_mesa_hash_table_u64_clear(struct hash_table_u64 *ht)
+_mesa_hash_table_u64_clear(struct hash_table_u64 *ht,
+                           void (*delete_function)(struct hash_entry *entry))
 {
    if (!ht)
       return;
 
-   _mesa_hash_table_clear(ht->table, _mesa_hash_table_u64_delete_key);
+   if (ht->deleted_key_data) {
+      if (delete_function) {
+         struct hash_table *table = ht->table;
+         struct hash_entry entry;
+
+         /* Create a fake entry for the delete function. */
+         if (sizeof(void *) == 8) {
+            entry.hash = table->key_hash_function(table->deleted_key);
+         } else {
+            struct hash_key_u64 _key = { .value = (uintptr_t)table->deleted_key };
+            entry.hash = table->key_hash_function(&_key);
+         }
+         entry.key = table->deleted_key;
+         entry.data = ht->deleted_key_data;
+
+         delete_function(&entry);
+      }
+      ht->deleted_key_data = NULL;
+   }
+
+   if (ht->freed_key_data) {
+      if (delete_function) {
+         struct hash_table *table = ht->table;
+         struct hash_entry entry;
+
+         /* Create a fake entry for the delete function. */
+         if (sizeof(void *) == 8) {
+            entry.hash = table->key_hash_function(uint_key(FREED_KEY_VALUE));
+         } else {
+            struct hash_key_u64 _key = { .value = (uintptr_t)FREED_KEY_VALUE };
+            entry.hash = table->key_hash_function(&_key);
+         }
+         entry.key = uint_key(FREED_KEY_VALUE);
+         entry.data = ht->freed_key_data;
+
+         delete_function(&entry);
+      }
+      ht->freed_key_data = NULL;
+   }
+
+   _mesa_hash_table_clear(ht->table, delete_function);
 }
 
 void
-_mesa_hash_table_u64_destroy(struct hash_table_u64 *ht)
+_mesa_hash_table_u64_destroy(struct hash_table_u64 *ht,
+                             void (*delete_function)(struct hash_entry *entry))
 {
    if (!ht)
       return;
 
-   _mesa_hash_table_u64_clear(ht);
-   _mesa_hash_table_destroy(ht->table, NULL);
+   _mesa_hash_table_u64_clear(ht, delete_function);
+   _mesa_hash_table_destroy(ht->table, delete_function);
    free(ht);
 }
